+++ src/domain/flow-system/src/services/flow/mod.rs
@@ -9,6 +9,8 @@
 
 mod flow_service;
 mod flow_service_event;
+mod flow_service_test_driver;
 
 pub use flow_service::*;
 pub use flow_service_event::*;
+pub use flow_service_test_driver::*;